fix (AVFoundation): set mPlayerVideoOutput’s delegate to nil when releasing
mPlayerVideoOutput’s delegate receives the outputSequenceWasFlushed message. Sometimes, when quickly releasing a lot of VideoOutputs, a message is received after the release resulting in a crash. Setting the delegate to nil discards this message.

A/B testing of this commit made it look like it was working as expected. No crashes with this line, some crashes without.